tg-bot/requirements.txt
python-telegram-bot[asyncio]>=20.0
httpx
python-dotenv 